% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{bind_attrs}
\alias{bind_attrs}
\title{Bind stars objects as attributes}
\usage{
bind_attrs(x, ...)
}
\arguments{
\item{x}{list of one or more stars objects}

\item{...}{ignored}
}
\value{
stars object
}
\description{
Bind stars objects as attributes
}
