% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{build_version}
\alias{build_version}
\title{Build a version string from components}
\usage{
build_version(major, minor, release, sep = ".")
}
\arguments{
\item{major}{string, like '3' or a named vector with (major, minor, [release])}

\item{minor}{string, like '13', ignored if major has length > 1}

\item{release}{string, optional like '002', ignored if major has length > 1}

\item{sep}{character, by default '.' but any single character will do}
}
\value{
version string like 'v3.13.002'
}
\description{
Versions have the format vMajor.Minor.Release
}
