% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{config_path}
\alias{config_path}
\title{Possibly retrieve a path from a configuration, but only if the input
is a configuration list}
\usage{
config_path(path, ...)
}
\arguments{
\item{path}{either a configuration list or a directory path}

\item{...}{other path segments to append to `path`}
}
\value{
a qualified path (but it might not exist)
}
\description{
Possibly retrieve a path from a configuration, but only if the input
is a configuration list
}
