% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopmodb-package.R
\docType{package}
\name{ecopmodb-package}
\alias{ecopmodb-package}
\alias{ecopmodb}
\title{ecopmodb}
\description{
Provides tools for navigating ecopmo prediction files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/ecopmodb}
}

}
\author{
\strong{Maintainer}: Tupper Ben \email{btupper@bigelow.org}

}
\keyword{internal}
