% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{make_path}
\alias{make_path}
\title{Create a path if it doesn't exist}
\usage{
make_path(path, recursive = TRUE, ...)
}
\arguments{
\item{path}{str one or more path descriptions}

\item{recursive}{logical, if TRUE then make intermediary paths as needed}

\item{...}{other arguments for `dir.create`}
}
\value{
the input path(s)
}
\description{
Create a path if it doesn't exist
}
