% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R, R/path.R
\name{parse_version}
\alias{parse_version}
\title{Parse a version string}
\usage{
parse_version(x = "v2.123.1", sep = ".")

parse_version(x = "v2.123.1", sep = ".")
}
\arguments{
\item{x}{version string to parse}

\item{sep}{character, by default '.' but any single character will do}
}
\value{
named character vector

named character vector
}
\description{
Versions have the format vMajor.Minor.Release

Versions have the format Major.Minor.Release
}
