% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{read_database}
\alias{read_database}
\alias{write_database}
\alias{append_database}
\title{Read, write and append a database}
\usage{
read_database(cfg, filename = "database")

write_database(x, cfg, filename = "database")

append_database(x, cfg, ...)
}
\arguments{
\item{cfg}{a configuration list}

\item{filename}{str the database file name}

\item{x}{a tabular database}

\item{...}{arguments passed to other functions such as `filename`}
}
\value{
a tabular database

the input tabular database
}
\description{
Read, write and append a database

Read, write and append a database

Append to a database
}
