% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{root_data_path}
\alias{root_data_path}
\title{Retrieve the root data path}
\usage{
root_data_path(..., root = get_root_data_path())
}
\arguments{
\item{...}{other path segments to append to the root}

\item{root}{chr the root data path}
}
\value{
the root data path
}
\description{
Retrieve the root data path
}
