% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{set_root_data_path}
\alias{set_root_data_path}
\alias{get_root_data_path}
\title{Set the root data path in a hidden file in the user's home
directory.}
\usage{
set_root_data_path(path = ".", filename = "~/.ecopmo", create = TRUE)

get_root_data_path(filename = "~/.ecopmo")
}
\arguments{
\item{path}{chr, the root data path}

\item{filename}{chr, the file where the path is stored}

\item{create}{logical, if TRUE create the path if it doesn't exist}
}
\value{
the input path
}
\description{
Set the root data path in a hidden file in the user's home
directory.
}
