% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{species_path}
\alias{species_path}
\alias{version_path}
\title{Generate a species path}
\usage{
species_path(cfg, ..., root = root_data_path())

version_path(cfg, ..., root = root_data_path())
}
\arguments{
\item{cfg}{a configuration list}

\item{...}{other path segments to add (by default none), 
these must appear before `root`}

\item{root}{str, the root data path}
}
\value{
the species or version data path
}
\description{
Generate a species path

Generate a version path from a version configuration
}
