% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_configuration}
\alias{write_configuration}
\alias{read_configuration}
\alias{list_configurations}
\title{Read, write and configurations}
\usage{
write_configuration(cfg, filename, root = root_data_path())

read_configuration(
  filename,
  species = NULL,
  version = NULL,
  root = root_data_path()
)

list_configurations(species = "testspecies", root = root_data_path())
}
\arguments{
\item{cfg}{list, configuration list}

\item{filename}{str the file description, not required but overrides the 
automatic construction of the output filename.  When reading a config, 
`filename` is ignored if species (and possibly version) are provided.}

\item{root}{the root data path}

\item{species}{str, the species name}

\item{version}{chr or NULL, if this and species are not NULL then search for the
version config in `root/species`}
}
\value{
configuration list (invisibly)

str, zero or more versions attached to this species
}
\description{
Read, write and configurations
}
