% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{build_database}
\alias{build_database}
\title{Build a database}
\usage{
build_database(path = ".", pattern = "^.*\\\\.tif$", save_db = FALSE)
}
\arguments{
\item{path}{a path to a database directory or a configuration list}

\item{pattern}{str, the regex pattern to use for finding files}

\item{save_db}{logical, if `TRUE` then save the database}
}
\value{
a tabular database
}
\description{
Build a database
}
