% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{config_path}
\alias{config_path}
\title{Retrieve a path from a configuration}
\usage{
config_path(cfg, ...)
}
\arguments{
\item{cfg}{a configuration list}

\item{...}{other path segments to append to the path}
}
\value{
a qualified path (but it might not exist)
}
\description{
Retrieve a path from a configuration
}
