% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{decompose_filename}
\alias{decompose_filename}
\alias{compose_filename}
\title{Compose, decompose and build prediction databases}
\usage{
decompose_filename(
  filename = c("1993-01-01_day_q000.tif", "1993-01-01_mon_mean.tif"),
  ext = ".tif"
)

compose_filename(x = decompose_filename(), path = ".", ext = ".tif")
}
\arguments{
\item{filename}{str, one or more filenames to decompose}

\item{ext}{the filename extension to remove or apply (with dot)}

\item{x}{a tabular database}

\item{path}{str the path to the data or a configuration list}
}
\value{
a tabular database with these elements
1. date Date, well, the date
2. per str, period such as 'day' or 'mon'
3. type str, descriptive where 'qNNN' is the N.NNth quantile

one or more filepaths
}
\description{
`decompose_filename` and `build_database` return a tabular database while
`compose_filename` returns a filename
}
