% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_ecopmo}
\alias{read_ecopmo}
\title{Read raster predictions}
\usage{
read_ecopmo(db, path, threshold = NULL)
}
\arguments{
\item{db}{data frame of file database}

\item{path}{chr, the path to the database}

\item{threshold}{num or NULL, if numeric then threshold the data to be 
greater than or equal to this value.}
}
\value{
stars object
}
\description{
Read raster predictions
}
